+++ src/Star.java
@@ -6,19 +6,19 @@ import com.badlogic.gdx.math.Vector3;
 
 public class Star extends SphericObject {
 
-    public Star(float radius, double mass) {
+    public Star(float radius, float mass) {
         super(new Vector3(0, 0, 0), radius, mass, new Velocity(new Vector3(0, 0, 0), 0));
     }
 
-    public Star(float radius, double mass, String name) {
+    public Star(float radius, float mass, String name) {
         super(new Vector3(0, 0, 0), radius, mass, new Velocity(new Vector3(0, 0, 0), 0), name);
     }
 
-    public Star(Vector3 position, float radius, double mass) {
+    public Star(Vector3 position, float radius, float mass) {
         super(position, radius, mass, new Velocity(new Vector3(0, 0, 0), 0));
     }
 
-    public Star(Vector3 position, float radius, double mass, String name) {
+    public Star(Vector3 position, float radius, float mass, String name) {
         super(position, radius, mass, new Velocity(new Vector3(0, 0, 0), 0), name);
     }
